+++ src/extension.ts
@@ -29,11 +29,9 @@ import { DefaultAWSStatusBar } from './shared/defaultStatusBar'
 import { EnvironmentVariables } from './shared/environmentVariables'
 import { ext } from './shared/extensionGlobals'
 import {
-    isDifferentVersion,
-    promptWelcome,
     safeGet,
-    setMostRecentVersion,
-    showQuickStartWebview
+    showQuickStartWebview,
+    toastNewUser
 } from './shared/extensionUtilities'
 import * as logFactory from './shared/logger'
 import { DefaultRegionProvider } from './shared/regions/defaultRegionProvider'
@@ -196,12 +194,7 @@ export async function activate(context: vscode.ExtensionContext) {
             '"'
         )
 
-        if (isDifferentVersion(context)) {
-            setMostRecentVersion(context)
-            // the welcome toast should be nonblocking.
-            // tslint:disable-next-line: no-floating-promises
-            promptWelcome()
-        }
+        toastNewUser(context)
 
         await resumeCreateNewSamApp(context)
     } catch (error) {